+++ service/persist/collection_token.go
@@ -12,7 +12,7 @@ import (
 // between collections and NFTS
 // This struct will only be used when updating or querying the database
 type CollectionTokenDB struct {
-	Version      NullInt64       `bson:"version" json:"version"` // schema version for this model
+	Version      NullInt32       `bson:"version" json:"version"` // schema version for this model
 	ID           DBID            `bson:"_id" json:"id" binding:"required"`
 	CreationTime CreationTime    `bson:"created_at" json:"created_at"`
 	Deleted      NullBool        `bson:"deleted" json:"-"`
@@ -34,7 +34,7 @@ type CollectionTokenDB struct {
 // This struct will always be decoded from a get database operation and will be used throughout
 // the application where CollectionDB does not apply
 type CollectionToken struct {
-	Version      NullInt64       `bson:"version"       json:"version"` // schema version for this model
+	Version      NullInt32       `bson:"version"       json:"version"` // schema version for this model
 	ID           DBID            `bson:"_id"           json:"id" binding:"required"`
 	CreationTime CreationTime    `bson:"created_at" json:"created_at"`
 	Deleted      NullBool        `bson:"deleted" json:"-"`
@@ -53,7 +53,7 @@ type CollectionToken struct {
 
 // TokenLayout defines the layout of a collection of tokens
 type TokenLayout struct {
-	Columns NullInt64 `bson:"columns" json:"columns"`
+	Columns NullInt32 `bson:"columns" json:"columns"`
 	// Padding         int   `bson:"padding" json:"padding"`
 	// WhitespaceAfter []int `bson:"whitespace_after" json:"whitespace_after"`
 }
